+++ Mythic_CLI/src/cmd/manager/managerInterface.go
@@ -66,6 +66,8 @@ type CLIManager interface {
 	PrintAllServices()
 	// ResetDatabase deletes the current database or volume
 	ResetDatabase(useVolume bool)
+	// ResetRabbitmq deletes the current rabbitmq storage or volume
+	ResetRabbitmq(useVolume bool)
 	// BackupDatabase saves a copy of the database to the specified path
 	BackupDatabase(backupPath string, useVolume bool) error
 	// RestoreDatabase restores a saved copy of the database from the specified path